pyproject.toml
[project]
name = "dict-ait-e"
version = "0.1.0"
description = "Voice transcription with interchangeable UIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "sounddevice",
    "soundfile",
    "openai>=1",
    "python-dotenv",
    "flask",
    "pydub",
    "pytest",
]

[project.optional-dependencies]
ui-gtk = ["PyGObject"]
ui-web = ["flask", "python-dotenv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
